Create 'flipProperty' module
Move the map of BiDi-equivalent properties, and the property-flipping
function, into their own private module.
